% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{camera_dotplot}
\alias{camera_dotplot}
\title{Create dot plots of cameraPR results by direction and contrast}
\usage{
camera_dotplot(x, n_path = 10)
}
\arguments{
\item{x}{data frame, containing results from cameraPR (with/without FDR filtering).
Should be combined to include all tests/contrasts/comparisons of interest in the experiment.
Must contain columns: \code{Direction}, \code{Contrast}, \code{NGenes}, \code{GeneSet}, \code{FDR}.}

\item{n_path}{integer, number of pathways to show for each direction (Up and Down).
Selects the top n based on the strongest enrichment (-log10(FDR)) for Up and Down separately.}
}
\value{
A named list of ggplot objects: \code{list(Up = <plot>, Down = <plot>)}.
You can further modify them with ggplot2 methods.
}
\description{
Create dot plots of cameraPR results by direction and contrast
}
