% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cmri_datatable}
\alias{cmri_datatable}
\title{Interactive data tables}
\usage{
cmri_datatable(df, digits = 2, ...)
}
\arguments{
\item{df}{Data frame}

\item{digits}{Number of digits to include in formatted numbers.}

\item{...}{Additional parameters to pass to \code{DT::datatable}}
}
\value{
Embedded HTML data table, \code{DT::datatable} object
}
\description{
Interactive data tables
}
\examples{

cmri_datatable(mtcars)



}
