% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmribio-package.R
\docType{package}
\name{cmribio-package}
\alias{cmribio}
\alias{cmribio-package}
\title{cmribio: Functions for Computational Bioinformatics And Reporting}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Collection of report templates and functions for standardized analyses performed by the BCB core biostatisticians and bioinformaticians.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jmvarberg.github.io/cmribio/}
}

}
\author{
\strong{Maintainer}: Joe Varberg \email{jmvarberg@cmh.edu}

}
\keyword{internal}
