% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{converse_gene_counts}
\alias{converse_gene_counts}
\title{Converse MNGC bulk RNAseq dataset}
\format{
## `converse_gene_counts`
A data frame with 78,317 rows (genes/features) and 14 columns (samples):
\describe{
  \item{gene_id}{ENSMBL ID}
  \item{transcript_id(s)}{transcript IDs associated with ENSMBL IDs}
  ...
}
}
\source{
<https://trace.ncbi.nlm.nih.gov/Traces/?view=study&acc=SRP549060>
}
\usage{
converse_gene_counts
}
\description{
RSEM merged gene counts from nf-core/rnaseq v. 3.19.0, from mouse study SRP549060
Report ...
}
\keyword{datasets}
