% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{converse_samplesheet}
\alias{converse_samplesheet}
\title{Converse nf-core/rnaseq sample sheet}
\format{
## `converse_samplesheet`
A data frame with 78,317 rows (genes/features) and 14 columns (samples):
\describe{
  \item{fastq_1, fastq_2}{paths to raw fastq files for processing}
  \item{strandedness}{param passed to nf-core/rnaseq to auto detect strandedness}
  \item{sample_title}{Grouping for cell type, replicate for experimental design}
  ...
}
}
\source{
<https://trace.ncbi.nlm.nih.gov/Traces/?view=study&acc=SRP549060>
}
\usage{
converse_samplesheet
}
\description{
Sample sheet with metadata from nf-core/rnasesq processing for converse_gene_counts
Report ...
}
\keyword{datasets}
