% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bcb_report.R
\name{create_bcb_report}
\alias{create_bcb_report}
\title{Generate a standard analysis report from template in the current directory.}
\usage{
create_bcb_report(file_name = NULL, ext_name = c("bcb-report", "bulk-rnaseq"))
}
\arguments{
\item{file_name}{Name for the Quarto report file.}

\item{ext_name}{Which template to use. One of c("bcb-report", "bulk-rnaseq"). Use "bcb-report" for general analyses, and "bulk-rnaseq" to perform DESeq2 analysis of RNAseq datasets.}
}
\value{
Copies template file into working directory.
}
\description{
Generate a standard analysis report from template in the current directory.
}
\examples{
if (interactive()) {
create_bcb_report(file_name = "BCB172_Analysis", ext_name = "bcb-report")
}
}
