% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_counts}
\alias{demo_counts}
\title{Example bulk RNAseq counts matrix}
\format{
A data frame with 12 columns/samples and 78,317 rows/genes/features:
}
\source{
Generated in \code{data-raw/demo_counts.R}.
}
\usage{
demo_counts
}
\description{
A gene x sample counts matrix as generated by nf-core/rnaseq with star_rsem
}
\details{
Used for demo purposes of '_targets_bulkRNA_DEseq2.R' demo pipeline for bulk RNAseq analysis. Generated using code in:
\code{data-raw/demo_counts.R}.
}
\examples{
# Basic usage
head(demo_counts)
dim(demo_counts)


}
\keyword{datasets}
