% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_samplesheet}
\alias{demo_samplesheet}
\title{Example bulk RNAseq sample sheet}
\format{
A data frame with 12 rows/samples and 5 columns of metadata:
}
\source{
Generated in \code{data-raw/demo_samplesheet.R}.
}
\usage{
demo_samplesheet
}
\description{
An example sample sheet for 'demo_counts' with metadata mapping sample names to group information.
}
\details{
Used for demo purposes of '_targets_bulkRNA_DEseq2.R' demo pipeline for bulk RNAseq analysis. Generated using code in:
\code{data-raw/demo_samplesheet.R}.
}
\examples{
# Basic usage
head(demo_samplesheet)
dim(demo_samplesheet)


}
\keyword{datasets}
