% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{get_genesets}
\alias{get_genesets}
\title{Pull Gene Sets of Interest from msigdbr package}
\usage{
get_genesets(
  species = c("human", "mouse"),
  sets = c("hallmark", "kegg", "wiki", "reactome", "go_bp", "go_cc", "go_mf")
)
}
\arguments{
\item{species}{"mouse" or "human"; default = "human"}

\item{sets}{which gene set collections to include, values = "hallmark", "kegg", "wiki", "reactome", "go_bp", "go_cc", "go_mf". Note, "kegg" not available for mouse.}
}
\value{
data frame of gene sets and their associated genes and information as output by msigdbr::msigdbr()
}
\description{
Pull Gene Sets of Interest from msigdbr package
}
\examples{
gs <- get_genesets(species = "human", sets = "hallmark")
head(gs)
}
