% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{init_proj}
\alias{init_proj}
\title{Initialize Project Directory}
\usage{
init_proj(dir = NULL, overwrite = c("ask", TRUE, FALSE))
}
\arguments{
\item{dir}{Path to parent directory to create subdirectories: data, documents, results, scripts and README.txt template}

\item{overwrite}{How to handle scenario when 'dir' provided or any of the subdirectories exist. Default "ask" prompts user for how to handle. TRUE overwrites existing, FALSE stops if 'dir' exists.}
}
\value{
Invisible file path to created 'dir'
}
\description{
Initialize Project Directory
}
\examples{
init_proj(dir = "./test_init", overwrite = "ask")

}
