% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_targets_templates}
\alias{list_targets_templates}
\title{List available targets pipeline templates}
\usage{
list_targets_templates()
}
\value{
Data frame with names and descriptions of templates available in the package, and they 'type' value to provide to cmribio::use_targets_template() to access.
}
\description{
List available targets pipeline templates
}
\examples{
list_targets_templates()

}
