% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{pal_bcb}
\alias{pal_bcb}
\title{BCB/CMRI Color Palettes}
\usage{
pal_bcb(alpha = 1)
}
\arguments{
\item{alpha}{Transparency level, real number between 0-1.}
}
\description{
Specific color palette based on CMRI brand guidlines.
}
\examples{
library(scales)
show_col(bcb_colors)
show_col(scales::alpha(bcb_colors, 0.5))


}
