% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway_analysis.R
\name{pathway_heatmap}
\alias{pathway_heatmap}
\title{Heatmap of pathway of interest}
\usage{
pathway_heatmap(
  expr_mat,
  gs_list,
  pathway_name,
  species = c("mouse", "human"),
  show_rowNames = FALSE
)
}
\arguments{
\item{expr_mat}{matrix, expression matrix with ENSEMBL IDs as row names}

\item{gs_list}{data frame, list of gene sets for species of interest}

\item{pathway_name}{character, name of pathway to use for heatmap.}

\item{species}{character, either "mouse" or "human"}

\item{show_rowNames}{boolean, whether to show row names on pheatmap::pheatmap output.}
}
\value{
pheatmap plot object
}
\description{
Heatmap of pathway of interest
}
