% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scale_color_bcb}
\alias{scale_color_bcb}
\alias{scale_fill_bcb}
\title{Functions to apply BCB Color Palettes}
\usage{
scale_color_bcb(alpha = 1, ...)

scale_fill_bcb(alpha = 1, ...)
}
\arguments{
\item{alpha}{Set transparency value for color/fill}

\item{...}{Additional arguments passed to ggplot2::discrete_scale}
}
\value{
ggplot2 scale object using the BCB discrete palette.
}
\description{
Functions to apply BCB Color Palettes
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt, color = as.factor(cyl))) + geom_point()
ggplot(mtcars, aes(mpg, wt, color = as.factor(cyl))) + geom_point() + scale_color_bcb()
}
