% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_deseq_targets}
\alias{use_deseq_targets}
\title{Use targets pipeline templates for bulk RNAseq analysis with DESeq2}
\usage{
use_deseq_targets(dir = NULL)
}
\arguments{
\item{dir}{default = NULL, which will use current working directory; path to directory to copy tempate files to}
}
\value{
copies three files: _targets.R (pipeline script), _targets_bulkRNA_DESeq2_functions.R (functions to run pipeline), and DESeq2_analysis_report.qmd (Quarto report template). Also makes copies of the demo counts matrix and sample sheet for testing the pipeline.
}
\description{
Use targets pipeline templates for bulk RNAseq analysis with DESeq2
}
