% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_targets_template}
\alias{use_targets_template}
\title{Use targets pipeline templates for bulk RNAseq analysis with DESeq2}
\usage{
use_targets_template(dir = NULL, type = c("deseq2"))
}
\arguments{
\item{dir}{default = NULL, which will use current working directory; path to directory to copy tempate files to}

\item{type}{character, choose which targets pipeline template to use. Available options include "deseq2".}
}
\value{
copies three files: _targets.R (pipeline script), _targets_bulkRNA_DESeq2_functions.R (functions to run pipeline), and DESeq2_analysis_report.qmd (Quarto report template). Also makes copies of the demo counts matrix and sample sheet for testing the pipeline.
}
\description{
Use targets pipeline templates for bulk RNAseq analysis with DESeq2
}
