% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grch38}
\alias{grch38}
\title{Annotation table for human using ENSMBL GRCh38 from stephenturner/annotables package.}
\format{
A data frame mouse annotation information.
}
\source{
Generated in \code{data-raw/grch38.R}.
}
\usage{
grch38
}
\description{
Annotation table for human using ENSMBL GRCh38 from stephenturner/annotables package.
}
\details{
Used for annotating datasets, swapping IDs between types etc.
\code{data-raw/grch38.R}.
}
\examples{
# Basic usage
head(grch38)
colnames(grch38)



}
\keyword{datasets}
