% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grcm38}
\alias{grcm38}
\title{Annotation table for mouse using ENSMBL GRCm38 from stephenturner/annotables package.}
\format{
A data frame mouse annotation information.
}
\source{
Generated in \code{data-raw/grcm38.R}.
}
\usage{
grcm38
}
\description{
Annotation table for mouse using ENSMBL GRCm38 from stephenturner/annotables package.
}
\details{
Used for annotating datasets, swapping IDs between types etc.
\code{data-raw/grcm38.R}.
}
\examples{
# Basic usage
head(grcm38)
colnames(grcm38)



}
\keyword{datasets}
