% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_targets_templates}
\alias{list_targets_templates}
\title{List available targets pipeline templates}
\usage{
list_targets_templates()
}
\value{
List of all targets templates included in the package
}
\description{
List available targets pipeline templates
}
