% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{use_targets_template}
\alias{use_targets_template}
\title{Use targets pipeline templates}
\usage{
use_targets_template(dir = NULL, type = c("deseq2"))
}
\arguments{
\item{dir}{default = NULL, which will use current working directory; path to directory to copy template files to}

\item{type}{character, choose which targets pipeline template to use. Available options include "deseq2".}
}
\value{
copies template files, functions, Quarto report templates and any demo data needed to run the pipeline into the working directory.
}
\description{
Use targets pipeline templates
}
